requirements.txt
# Test Dependencies
pynose
pinocchio
factory-boy

# Code Coverage
coverage

requests

python-dotenv
